armor: fix potential memory leak of pgp_plain.
If realloc() fails, it will return NULL, which we promptly wrote to the
pgp_plain pointer, upon trying to free() it afterwards. Not a problem in
and of itself, the problem being that realloc() does not free() the
original pointer if it fails, so this was a memory leak.
Caught by clang-analyze.